requirements.txt
fastapi
uvicorn
gunicorn
ortools
requests
python-dotenv
